requirements.txt
fastapi == 0.115.6
uvicorn == 0.34.0
scikit-learn == 1.3.2
xgboost == 2.1.0
imbalanced-learn == 0.13.0
python-dotenv == 1.0.1
matplotlib == 3.10.0
seaborn == 0.13.2
python-multipart == 0.0.20
numpy == 1.26.4
opencv-python == 4.9.0.80
scikit-image == 0.22.0
scipy == 1.12.0
Pillow == 10.2.0
joblib == 1.3.2
tensorflow == 2.15.0
protobuf == 3.20.3
